+++ PeriodicalsServlet/src/main/java/ua/training/model/dao/mapper/StudentMapper.java
@@ -19,7 +19,7 @@ public class StudentMapper implements ObjectMapper<Student> {
     }
 
     @Override
-    public Student makeUnique(Map<Integer, Student> cache,
+    public Student makeUnique(Map<Long, Student> cache,
                               Student student) {
         cache.putIfAbsent(student.getId(), student);
         return cache.get(student.getId());
